wms-gui/requirements.txt
requests~=2.28.2
streamlit~=1.21.0
PyYAML~=6.0